+++ sourcespec/ssp_grid_sampling.py
@@ -15,6 +15,7 @@ The class provides optimal solutions, uncertainties and plotting methods.
 """
 import os
 import numpy as np
+import matplotlib.pyplot as plt
 import logging
 logger = logging.getLogger(__name__.split('.')[-1])
 
@@ -27,18 +28,24 @@ class GridSampling():
     The class provides optimal solutions, uncertainties and plotting methods.
     """
 
-    def __init__(self, misfit_func, bounds, nsteps, sampling_mode):
+    def __init__(self, misfit_func, bounds, nsteps, sampling_mode,
+                 params_name, params_unit):
         """
         Init grid sampling.
 
         bounds : sequence of (min, max) pairs for each dimension.
         nsteps : number of grid steps for each dimension.
         sampling_mode : sequence of 'lin' or 'log' for each diemesion.
+        params_name : sequence of parameter names (str).
+        params_name : sequence of parameter units (str).
         """
         self.misfit_func = misfit_func
         self.bounds = bounds
+        self.truebounds = None
         self.nsteps = nsteps
         self.sampling_mode = sampling_mode
+        self.params_name = params_name
+        self.params_unit = params_unit
         self.misfit = None
         self._values = None
         self._min_idx = None
@@ -48,6 +55,7 @@ class GridSampling():
         if self._values is not None:
             return self._values
         values = []
+        self.truebounds = []
         for bds, ns, mode in zip(self.bounds, self.nsteps, self.sampling_mode):
             if None in bds:
                 msg = 'All parameters must be bounded for grid sampling'
@@ -59,6 +67,7 @@ class GridSampling():
                 values.append(np.logspace(*bds, ns))
             else:
                 values.append(np.linspace(*bds, ns))
+            self.truebounds.append(bds)
         self._values = np.meshgrid(*values, indexing='ij')
         return self._values
 
@@ -85,9 +94,8 @@ class GridSampling():
         mf = np.vectorize(mf)
         self.misfit = mf(*self.values)
 
-    def plot_conditional_misfit(self, config, params_name, params_unit, label):
+    def plot_conditional_misfit(self, config, label):
         """Plot conditional misfit for each parameter."""
-        import matplotlib.pyplot as plt
         ndim = self.misfit.ndim
         fig, ax = plt.subplots(ndim, 1, figsize=(9, 9), dpi=300)
         for dim in range(ndim):
@@ -111,8 +119,8 @@ class GridSampling():
                 transform=ax[dim].get_xaxis_transform())
             if self.sampling_mode[dim] == 'log':
                 ax[dim].set_xscale('log')
-            xlabel = params_name[dim]
-            unit = params_unit[dim]
+            xlabel = self.params_name[dim]
+            unit = self.params_unit[dim]
             if unit:
                 xlabel += ' ({})'.format(unit)
             ax[dim].set_xlabel(xlabel)
@@ -135,4 +143,99 @@ class GridSampling():
             fig.savefig(figfile, bbox_inches='tight')
             logger.info(
                 '{}: conditional misfit plot saved to: {}'.format(
+                    label, figfile))
+
+    def plot_misfit_2d(self, config, plot_par_idx, label):
+        """Plot a 2D misfit map."""
+        # Move axes to keep at the end
+        if plot_par_idx[0] < plot_par_idx[1]:
+            end_idx = (-2, -1)
+        else:
+            end_idx = (-1, -2)
+        mm = np.moveaxis(self.misfit, plot_par_idx, end_idx)
+        # Find the index to extract
+        idx = tuple(
+            v for n, v in enumerate(self.min_idx) if n not in plot_par_idx)
+        # extract a 2D misfit map
+        mm = mm[idx]
+        # set extent for imshow
+        extent = []
+        for n, bds in enumerate(self.truebounds):
+            if n not in plot_par_idx:
+                continue
+            extent += list(bds)
+        params_opt = []
+        for popt, mode in zip(self.params_opt, self.sampling_mode):
+            if mode == 'log':
+                popt = np.log10(popt)
+            params_opt.append(popt)
+        # extract parameter info only for the two selected parameters
+        params_opt = [
+            p for n, p in enumerate(params_opt) if n in plot_par_idx]
+        params_name = [
+            p for n, p in enumerate(self.params_name) if n in plot_par_idx]
+        params_unit = [
+            p for n, p in enumerate(self.params_unit) if n in plot_par_idx]
+        sampling_mode = [
+            p for n, p in enumerate(self.sampling_mode) if n in plot_par_idx]
+        fig, ax = plt.subplots(1, 1, figsize=(9, 8), dpi=300)
+        # imshow: saturate scale at 2 times the minimum
+        mmap = ax.imshow(
+            mm.T, vmax=2*mm.min(), origin='lower', cmap='viridis',
+            extent=extent, aspect='auto'
+        )
+        ax.scatter(*params_opt, facecolors='none', edgecolors='w')
+        ax.set_title(label)
+        xlabel = params_name[0]
+        unit = params_unit[0]
+        if unit:
+            xlabel += ' ({})'.format(unit)
+        if sampling_mode[0] == 'log':
+            # the grid is plotted by imshow() in linear scale,
+            # so we create a fake xaxis with logscale
+            ax.xaxis.set_visible(False)
+            ax2 = ax.twiny()
+            ax2.xaxis.set_label_position('bottom')
+            ax2.xaxis.set_ticks_position('bottom')
+            ax2.set_xlim(10**extent[0], 10**extent[1])
+            ax2.set_xscale('log')
+            ax2.set_xlabel(xlabel)
+        else:
+            ax.set_xlabel(xlabel)
+        ylabel = params_name[1]
+        unit = params_unit[1]
+        if unit:
+            ylabel += ' ({})'.format(unit)
+        if sampling_mode[1] == 'log':
+            # the grid is plotted by imshow() in linear scale,
+            # so we create a fake yaxis with logscale
+            ax.yaxis.set_visible(False)
+            ax2 = ax.twinx()
+            ax2.yaxis.set_label_position('left')
+            ax2.yaxis.set_ticks_position('left')
+            ax2.set_ylim(10**extent[2], 10**extent[3])
+            ax2.set_yscale('log')
+            ax2.set_ylabel(ylabel)
+        else:
+            ax.set_ylabel(ylabel)
+        cbar = plt.colorbar(mmap, ax=ax, extend='max')
+        cbar.set_label('misfit')
+        outdir = os.path.join(config.options.outdir, 'misfit')
+        if not os.path.exists(outdir):
+            os.makedirs(outdir)
+        evid = config.hypo.evid
+        figfile_base = os.path.join(outdir, evid)
+        suffix = '{}-{}'.format(*params_name)
+        figfile_base += '.misfit_{}_{}.'.format(
+            suffix, label.replace(' ', '_'))
+        fmt = config.PLOT_SAVE_FORMAT
+        if fmt == 'pdf_multipage':
+            fmt = 'pdf'
+        figfile = figfile_base + fmt
+        if config.PLOT_SHOW:
+            plt.show()
+        if config.PLOT_SAVE:
+            fig.savefig(figfile, bbox_inches='tight')
+            logger.info(
+                '{}: misfit map saved to: {}'.format(
                     label, figfile))